test: update file_sums, hdcd20.wav, hdcd24.wav
As I don't have any legit 20 or 24-bit HDCD-encoded PCM to test, I've
simply copied hdcd.wav into the top 16-bits of 24-bit PCM, and then
duplicated the 16-bit LSB into the LSB(bit 0) of the 24-bit PCM for
hdcd24.wav, and the 16-bit LSB into bit 4 of the 24-bit PCM for
hdcd20.wav. This only serves to test that hdcd-detect is scanning the
right bits for a given bit depth.

I don't have enough information to fully decode 20 or 24-bit HDCD at
this time.

Signed-off-by: Burt P <pburt0@gmail.com>