tool: calculate expected output size for piped WAV output
When using a non-seekable stream, the data length can't be set on
close, so calculate an expected data size and write that when writing
the WAV header. If the stream is seekable, it will still be replaced
at close with the actual value.

This should now work as expected:
flac -dc something16.flac | hdcd-detect -k | flac -o something24.flac -

Signed-off-by: Burt P <pburt0@gmail.com>